Check that the genome file exists before running prokka. If not, write an error message

+++ scripts/prokka_array.sh
@@ -7,6 +7,7 @@ dateinit=$4
 force=$5
 cpus=$6
 
+
 module load hmmer/3.1b1 aragorn/1.2.36 barrnap minced/0.1.6 blast+ prodigal infernal/1.1 ncbi_toolbox/20151127 signalp
 module load prokka/1.11
 
@@ -18,13 +19,18 @@ nbcont=${vars[2]}
 logfile=log-$gemname-$dateinit.out
 errfile=err-$gemname-$dateinit.err
 
+# Check that the oriname file exists. Otherwise, put a message in errfile and exit.
+if [ ! -f $oriname-gembase.fna ]; then
+	echo $oriname-gembase.fna "(from original file '$oriname') not found. Annotation and formatting pipeline cannot run." | tee -a $errfile -a $logfile
+	exit 1
+fi
+
 if [ ! $gemname == "gembase_name" ]; then
 	echo "--------------------------" > $logfile
 	echo "*" `date +"%d/%m/%y - %T"` "prokka: task" ${SGE_TASK_ID}: $gemname $oriname $nbcont "contigs" >> $logfile
 	# If prokka res folder already exists, write WARNING because it will use already generated results
 	if [ -d $oriname-prokka11Res ] && [ -z $force ]; then
-		echo "/!\\" `date +"%d/%m/%y - %T"` $gemname $oriname  "WARNING: prokka results folder already exists. Prokka will not run again, formatting step will use already generated results of Prokka in " $oriname-prokka11Res. "If you want to re-run prokka, first remove this result folder, or use '-f' option if you want to rerun prokka for all genomes." >> $logfile
-		echo "/!\\" `date +"%d/%m/%y - %T"` $gemname $oriname  "WARNING: prokka results folder already exists. Prokka will not run again, formatting step will use already generated results of Prokka in " $oriname-prokka11Res. "If you want to re-run prokka, first remove this result folder, or use '-f' option if you want to rerun prokka for all genomes." >> $errfile
+		echo "/!\\" `date +"%d/%m/%y - %T"` $gemname $oriname  "WARNING: prokka results folder already exists. Prokka will not run again, formatting step will use already generated results of Prokka in " $oriname-prokka11Res. "If you want to re-run prokka, first remove this result folder, or use '-f' option if you want to rerun prokka for all genomes." | tee -a $errfile -a $logfile
 	fi
 	prokka --outdir $oriname-prokka11Res --cpus $cpus $force --prefix $gemname $oriname"-gembase.fna"
 	# --centre to generate clean contig names (length <= 20)
